infra/requirements.txt
aws-cdk-lib>=2.150.0
constructs>=10.0.0
aws-cdk.aws-bedrock-agentcore-alpha
aws-cdk.aws-bedrock-alpha
